two of 4 tasks done some improvments

+++ src/pages/users/UsersPage.tsx
@@ -4,10 +4,10 @@ import Popup from 'reactjs-popup';
 import Input from "../../components/Input";
 import Select from "react-select";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {faPenToSquare, faTrash} from "@fortawesome/free-solid-svg-icons";
+import {faClose, faPenToSquare, faTrash} from "@fortawesome/free-solid-svg-icons";
 import {createUser, deleteUser, getUsers, User, UserRequest} from "./UsersAPI";
 import {useNavigate} from "react-router-dom";
-import {Form, FormContainer, InputWrapper, MainContainer, TableContainer} from "../Styled";
+import {CloseButton, Form, FormContainer, InputWrapper, MainContainer, TableContainer} from "../Styled";
 import {getRoles, Role} from "../roles/RolesAPI";
 
 
@@ -16,22 +16,28 @@ export default () => {
     const navigator = useNavigate();
     const [users, setUsers] = useState<User[]>([]);
     const [roles, setRoles] = useState<Role[]>([]);
+    let currentUser = localStorage.getItem("access_token");
+    if (!currentUser) currentUser = "";
+    let hasDeletePermission = JSON.parse(atob(currentUser.split(".")[1])).roles.includes('ROLE_DELETE_USER');
+    let hasUpdatePermission = JSON.parse(atob(currentUser.split(".")[1])).roles.includes('ROLE_UPDATE_USER');
+    let hasCreatePermission = JSON.parse(atob(currentUser.split(".")[1])).roles.includes('ROLE_CREATE_USER');
     useEffect(
         () => {
-            updateUsers()
+            updateUsers().then()
             getRoles().then(content => setRoles(content))
-        }
-        , []
+        }, []
     );
 
     function removeUser(user: User) {
-        deleteUser(user.username).then(() => {
-                let newUsers = users.filter(function (u) {
-                    return u !== user
-                })
-                setUsers(newUsers)
-            }
-        )
+        if (window.confirm(`are you sure about removing user ${user.username} ?`)) {
+            deleteUser(user.username).then(() => {
+                    let newUsers = users.filter(function (u) {
+                        return u !== user
+                    })
+                    setUsers(newUsers)
+                }
+            )
+        }
     }
 
     const updateUsers = () => getUsers().then(content => {
@@ -42,15 +48,21 @@ export default () => {
 
     return (
         <MainContainer>
+            {hasCreatePermission &&
             <Popup key={"create_user"}
                    trigger={<button className={"create-button"}>ساخت</button>}
                    position="center center"
                    modal nested closeOnEscape>
                 {(close: () => any) => (
                     <FormContainer>
+                        <CloseButton>
+                            <FontAwesomeIcon cursor={"pointer"} icon={faClose} height="100%" width="100%"
+                                             onClick={() => close()}/>
+                        </CloseButton>
+
                         <Form onSubmit={handleSubmit((user) => {
                             createUser(user).then(() => {
-                                getUsers().then(() => updateUsers())
+                                updateUsers().then()
                                 close()
                             });
                         })}>
@@ -69,7 +81,7 @@ export default () => {
                             <br/>
                             <Select options={roles.map(role => {
                                 return {
-                                    value: role.name,
+                                    value: role.id,
                                     label: role.name,
                                 }
                             })}
@@ -77,16 +89,16 @@ export default () => {
                                     isClearable={false}
                                     placeholder={"نقش"}
                                     onChange={option => {
-                                        setValue("roleName", option != null ? option.value : "")
+                                        setValue("roleId", option != null ? option.value : 0)
                                     }}
                             />
                             <br/>
                             <Input type="submit" value="ساخت"/>
-                            <button onClick={() => close()}> بستن </button>
                         </Form>
                     </FormContainer>
                 )}
             </Popup>
+            }
             <TableContainer>
                 <table>
                     <thead>
@@ -95,8 +107,8 @@ export default () => {
                         <td>نام</td>
                         <td>نام خانوادگی</td>
                         <td>نقش</td>
-                        <td>ویرایش</td>
-                        <td>حذف</td>
+                        {hasUpdatePermission && <td>ویرایش</td>}
+                        {hasDeletePermission && <td>حذف</td>}
                     </tr>
                     </thead>
                     <tbody>
@@ -108,6 +120,7 @@ export default () => {
                                     <td>{user.firstName}</td>
                                     <td>{user.lastName}</td>
                                     <td>{user.roleName}</td>
+                                    {hasUpdatePermission &&
                                     <td>
                                         <FontAwesomeIcon
                                             cursor={"pointer"}
@@ -115,6 +128,8 @@ export default () => {
                                             onClick={() => navigator("/users/" + user.username)}
                                         />
                                     </td>
+                                    }
+                                    {hasDeletePermission &&
                                     <td>
                                         <FontAwesomeIcon
                                             cursor={"pointer"}
@@ -122,6 +137,7 @@ export default () => {
                                             onClick={() => removeUser(user)}
                                         />
                                     </td>
+                                    }
                                 </tr>
                             )
                         )
